starting point
This platform is a contract exchange  allowing users to create,  and participate on smart contracts. It uses XRP ledger for secure and fast execution and settlement, it  offers templates for various use cases. A contract rating system and stats adds transparency, making it a reliable solution for businesses and individuals looking for take part.